use bigint for as number as private 32 bit as number will sometime enter routing tables, and the signed 32 bit integer of postgres cannot store it

+++ data/schema.sql
@@ -26,7 +26,7 @@ CREATE VIEW managed_routes_view AS
 -- consider expanding this with start/end timestamp later
 CREATE TABLE routes (
     route       cidr                NOT NULL,
-    asn         integer             NOT NULL,
+    asn         bigint              NOT NULL,
     source_id   integer             NOT NULL REFERENCES sources(id),
 
     -- The unique constraint may not be the best from an efficiancy point of view,
